requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
selenium>=4.10.0
python-dotenv>=1.0.0
